feat(server): POST /extract, POST /import, GET /export endpoints (#604 #605 #606)
Move Extractor from uteke-cli to uteke-core for server reuse.
Add three new HTTP endpoints to uteke-serve:

- POST /extract: LLM fact extraction + auto-store (1MB limit)
- POST /import: JSONL import with re-embedding (5MB limit)
- GET /export: JSONL export (optional ?namespace= filter)

Extraction reads [extraction] config from uteke.toml.
API key never accepted from request body (config/env only).
Write token required for all three endpoints.

Closes #604, #605, #606
